Tests descriptions added. Tests ran. Successful tests: T002_getArea_4and6_24 T003_setLength_55 T004_setLength_100 T006_setWidth_45 T008_setWidth_0
Tests failed + reasons:
T001_getPerimeter_4and6_20 failed because the formula of the perimeter is (P = 2a+2b), but the formula used in getPerimeter in main.c is (P = 2a + b).

T005_setLength_0 failed because in function setLength in main.c value of 0 is included in the range of accepted numbers, but according to the assignment it should be excluded.

T007_setWidth_100 failed because in function setWidth in main.c value of 100 is included in the range of accepted numbers, but according to the assignment it should be excluded.